+++ src/Authenticator/Contracts/Authenticator.php
@@ -11,22 +11,22 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Cog\YouTrack\Rest\Authorizer\Contracts;
+namespace Cog\YouTrack\Rest\Authenticator\Contracts;
 
 use Cog\YouTrack\Rest\Client\Contracts\Client as ClientContract;
 
 /**
  * Interface Authorizer.
  *
- * @package Cog\YouTrack\Rest\Authorizer\Contracts
+ * @package Cog\YouTrack\Rest\Authenticator\Contracts
  */
-interface Authorizer
+interface Authenticator
 {
     /**
-     * Returns authorization headers.
+     * Authenticate API Client.
      *
      * @param \Cog\YouTrack\Rest\Client\Contracts\Client $client
      * @return void
      */
-    public function appendHeadersTo(ClientContract $client): void;
+    public function authenticate(ClientContract $client): void;
 }